move RemixLink customisation in custom-components
