+++ packages/react-sdk/src/app/custom-components/rich-text-link.tsx
@@ -1,0 +1,13 @@
+import { forwardRef } from "react";
+import {
+  type RichTextLinkProps,
+  type RichTextLinkRef,
+  RichTextLink,
+} from "../../components/rich-text-link";
+import { renderRemixLink } from "./shared/remix-link";
+
+export const Component = forwardRef<RichTextLinkRef, RichTextLinkProps>(
+  ({ href = "", ...props }, ref) => renderRemixLink(href, props, ref)
+);
+
+Component.displayName = RichTextLink.displayName;
