+++ packages/react-sdk/src/components/link.tsx
@@ -1,29 +1,13 @@
-import { Link as RemixLink } from "@remix-run/react";
 import { forwardRef, type ElementRef, type ComponentProps } from "react";
 
-const isAbsoluteUrl = (href: string) => {
-  // http:, mailto:, tel:, etc.
-  return /^[a-z]+:/i.test(href);
-};
-
-export const renderLink = (
-  href: string,
-  props: Omit<ComponentProps<"a">, "href">,
-  ref: React.ForwardedRef<HTMLAnchorElement>
-) => {
-  return isAbsoluteUrl(href) ? (
-    <a {...props} href={href} ref={ref} />
-  ) : (
-    <RemixLink {...props} to={href} ref={ref} />
-  );
-};
-
-type LinkProps = Omit<ComponentProps<"a">, "href"> & {
+export type LinkProps = Omit<ComponentProps<"a">, "href"> & {
   href?: string;
 };
 
-export const Link = forwardRef<ElementRef<"a">, LinkProps>(
-  ({ href = "", ...props }, ref) => renderLink(href, props, ref)
+export type LinkRef = ElementRef<"a">;
+
+export const Link = forwardRef<LinkRef, LinkProps>(
+  ({ href = "", ...props }, ref) => <a {...props} href={href} ref={ref} />
 );
 
 Link.displayName = "Link";
